Improvements and fixes in newquest helper script.
